Restructure models module
