+++ fixmyapp/models/__init__.py
@@ -1,0 +1,13 @@
+# fmt: off
+from .bike_stands import BikeStands
+from .gastro_signup import GastroSignup, get_upload_path
+from .like import Like
+from .photo import Photo
+from .playstreet_signup import PlaystreetSignup
+from .profile import Profile
+from .project import Project
+from .question import Question
+from .report import Report
+from .section import Section
+from .section_details import SectionDetails
+# fmt: on
