+++ fixmyapp/models/like.py
@@ -1,0 +1,14 @@
+from django.contrib.auth import get_user_model
+from django.contrib.contenttypes.fields import GenericForeignKey
+from django.contrib.contenttypes.models import ContentType
+from django.contrib.gis.db import models
+
+from .base_model import BaseModel
+
+
+class Like(BaseModel):
+    content_object = GenericForeignKey('content_type', 'object_id')
+    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
+    object_id = models.PositiveIntegerField()
+    user = models.ForeignKey(get_user_model(), on_delete=models.CASCADE)
+
