+++ fixmyapp/models/section.py
@@ -1,0 +1,35 @@
+from django.contrib.gis.db import models
+from django.utils.translation import gettext_lazy as _
+
+from .base_model import BaseModel
+
+
+class Section(BaseModel):
+    street_name = models.CharField(_('street name'), max_length=100)
+    suffix = models.CharField(_('suffix'), blank=True, null=True, max_length=3)
+    borough = models.CharField(_('borough'), blank=True, null=True, max_length=255)
+    street_category = models.PositiveSmallIntegerField(_('street category'), null=True)
+    geometry = models.MultiLineStringField(_('geometry'), srid=4326, null=True)
+
+    class Meta:
+        verbose_name = _('section')
+        verbose_name_plural = _('sections')
+
+    def velocity_index(self):
+        if len(self.details.all()) > 0:
+            return sum(d.velocity_index() for d in self.details.all()) / len(
+                self.details.all()
+            )
+        else:
+            return 0
+
+    def safety_index(self):
+        if len(self.details.all()) > 0:
+            return sum(d.safety_index() for d in self.details.all()) / len(
+                self.details.all()
+            )
+        else:
+            return 0
+
+    def __str__(self):
+        return '{} ({})'.format(self.street_name, self.id)
