+++ survey/models.py
@@ -1,7 +1,7 @@
 from django.contrib.postgres.fields import JSONField
 from django.db import models
 from django.utils.translation import gettext_lazy as _
-from fixmyapp.models import BaseModel
+from fixmyapp.models.base_model import BaseModel
 import numpy
 
 
@@ -42,23 +42,21 @@ class Scene(BaseModel):
 
     def __str__(self):
         return '{:02}_{}_{}_{}'.format(
-            self.project, self.experiment, self.perspective, self.number)
+            self.project, self.experiment, self.perspective, self.number
+        )
 
     @classmethod
     def find_by_scene_id(cls, scene_id):
         parts = scene_id.split('_')
         return cls.objects.get(
-            project=parts[0],
-            experiment=parts[1],
-            perspective=parts[2],
-            number=parts[3]
+            project=parts[0], experiment=parts[1], perspective=parts[2], number=parts[3]
         )
 
     @classmethod
     def random_group(cls, perspective, project, size):
         scenes = cls.objects.filter(perspective=perspective, project=project)
         d = sum(s.weight for s in scenes)
-        p = [s.weight/d for s in scenes]
+        p = [s.weight / d for s in scenes]
         return numpy.random.choice(scenes, size, replace=False, p=p)
 
 
@@ -75,10 +73,10 @@ class Session(BaseModel):
 class Rating(BaseModel):
     duration = models.PositiveIntegerField(_('duration'), null=True)
     rating = models.PositiveSmallIntegerField(_('rating'), null=True)
-    scene = models.ForeignKey(
-        Scene, on_delete=models.SET_NULL, null=True)
+    scene = models.ForeignKey(Scene, on_delete=models.SET_NULL, null=True)
     session = models.ForeignKey(
-        Session, related_name='ratings', on_delete=models.CASCADE)
+        Session, related_name='ratings', on_delete=models.CASCADE
+    )
 
     class Meta:
         verbose_name = _('rating')
